Fix compile warning.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5720 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/irrString.h
@@ -1269,7 +1269,7 @@ public:
 	string<T,TAlloc>& eraseTrailingFloatZeros(char decimalPoint='.')
 	{
 		s32 i=findLastCharNotInList("0", 1);
-		if ( i > 0 && i<used-2 )	// non 0 must be found and not last char (also used is at least 2 when i > 0)
+		if ( i > 0 && (u32)i < used-2 )	// non 0 must be found and not last char (also used is at least 2 when i > 0)
 		{
 			u32 eraseStart=i+1;
 			u32 dot=0;
